+++ CHANGELOG.md
@@ -17,6 +17,7 @@ incremented for features.
 
 * lang: Add fallback functions ([#457](https://github.com/project-serum/anchor/pull/457)).
 * lang: Add feature flag for using the old state account discriminator. This is a temporary flag for those with programs built prior to v0.7.0 but want to use the latest Anchor version. Expect this to be removed in a future version ([#446](https://github.com/project-serum/anchor/pull/446)).
+* lang: Add generic support to Accounts ([#496](https://github.com/project-serum/anchor/pull/496)).
 
 ### Breaking Changes